pyproject.toml
[project]
name = "apple-calendar-mcp"
version = "0.1.0"
description = "MCP server for Apple Calendar via EventKit"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "mcp[cli]>=1.12",
    "pyobjc-framework-EventKit>=11.0; sys_platform == 'darwin'",
]

[project.scripts]
apple-calendar-mcp = "apple_calendar_mcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/apple_calendar_mcp"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.15.5",
]
